requirements.txt
pymupdf==1.26.6
pytesseract==0.3.13
pillow==12.0.0
regex==2025.11.3
pandas==2.3.3
numpy==2.2.6
scikit-learn==1.7.2
tqdm==4.67.1
